split exclusions code into separate modules
